BASIC WORLD GENERATION

+++ BasicVoxelEngine/Vector3.cs
@@ -85,7 +85,7 @@ namespace BasicVoxelEngine
         {
             return (X.GetHashCode() << 2) ^ (Y.GetHashCode() << 1) ^ Z.GetHashCode();
         }
-        public float TaxicabLength => Math.Abs(X) + Math.Abs(Y) + Math.Abs(Z);
+        public int TaxicabLength => Math.Abs(X) + Math.Abs(Y) + Math.Abs(Z);
         public Vector3I Dot(Vector3I vector) => new Vector3I(X * vector.X, Y * vector.Y, Z * vector.Z);
         public static Vector3I operator+(Vector3I vector1, Vector3I vector2) => new Vector3I(vector1.X + vector2.X, vector1.Y + vector2.Y, vector1.Z + vector2.Z);
         public static Vector3I operator+(Vector3I vector, int scalar) => vector + new Vector3I(scalar);
